requirements.txt
playwright
pytest
pytest-playwright
allure-pytest
allure-python-commons
pandas
openpyxl
requests
